Refactor BTC wallet balance calculation by extracting `calculateBtcWalletBalance` into a dedicated service, simplifying listener logic.
