+++ src/frontend/src/btc/services/btc-listener.services.ts
@@ -1,11 +1,8 @@
-import { loadBtcPendingSentTransactions } from '$btc/services/btc-pending-sent-transactions.services';
+import { calculateBtcWalletBalance } from '$btc/services/wallet-btc.service';
 import { btcTransactionsStore } from '$btc/stores/btc-transactions.store';
 import type { BtcTransactionUi } from '$btc/types/btc';
 import type { BtcPostMessageDataResponseWallet } from '$btc/types/btc-post-message';
-import { getBtcSourceAddress } from '$btc/utils/btc-address.utils';
-import { getBtcWalletBalance, mapTokenIdToNetworkId } from '$icp/utils/btc.utils';
 import { getIdbBtcTransactions } from '$lib/api/idb-transactions.api';
-import { authIdentity } from '$lib/derived/auth.derived';
 import { syncWalletFromIdbCache } from '$lib/services/listener.services';
 import { balancesStore } from '$lib/stores/balances.store';
 import { i18n } from '$lib/stores/i18n.store';
@@ -45,36 +42,14 @@ export const syncWallet = async ({
 	}
 	if (nonNullish(balance)) {
 		/*
-		 * Balance calculation is performed here in the main thread rather than in the worker (btc-wallet.scheduler.ts)
+		 * The calance calculation is performed here in the main thread rather than in the worker (btc-wallet.scheduler.ts)
 		 * because the pending transactions store (btcPendingSentTransactionsStore) is not accessible from the worker context.
 		 * The worker provides the confirmed balance from the Bitcoin canister, and we calculate the structured
 		 * balance (confirmed, unconfirmed, total) using newTransactions data to determine confirmation states.
 		 */
-		const identity = get(authIdentity);
-
-		// Get networkId from tokenId (same way as SendContext derives it from token)
-		const networkId = mapTokenIdToNetworkId(tokenId);
-
-		// Get the source address using the same logic as other components (BtcConvertTokenWizard, etc.)
-		const sourceAddress = getBtcSourceAddress(networkId);
-
-		// Wait for pending transactions to be loaded before calculating balance
-		await loadBtcPendingSentTransactions({
-			identity,
-			networkId,
-			address: sourceAddress
-		});
-
-		// Get transactions directly from the store instead of using parsed worker data
-		const storeData = get(btcTransactionsStore);
-		const storeTransactions = nonNullish(storeData) ? (storeData[tokenId] ?? null) : null;
-
-		// Calculate the structured balance using parsed transactions
-		// Use sourceAddress to ensure consistency with the address used for pending transactions
-		const btcWalletBalance = getBtcWalletBalance({
-			address: sourceAddress,
+		const btcWalletBalance = await calculateBtcWalletBalance({
 			balance,
-			providerTransactions: storeTransactions
+			tokenId
 		});
 
 		balancesStore.set({
